add code for plotting, fix types in C++

+++ OU4/ou4_2_fib/heltal.cpp
@@ -1,15 +1,16 @@
 #include <cstdlib>
+#include <stdint.h>
 // Integer class 
 
 class Heltal{
 	public:
 		Heltal(int);
-		int fib();
+		uint64_t fib();
 		int get();
 		void set(int);
 	private:
 		int val;
-		int calc_fib(int);
+		uint64_t calc_fib(int);
 	};
  
 Heltal::Heltal(int n){
@@ -24,14 +25,14 @@ void Heltal::set(int n){
 	val = n;
 	}
 
-int Heltal::calc_fib(int n) {
+uint64_t Heltal::calc_fib(int n) {
     if (n <= 1)
         return n;
     else 
         return (calc_fib(n-1) + calc_fib(n-2));    
     }
 
-int Heltal::fib() {
+uint64_t Heltal::fib() {
     return calc_fib(val);
     }
 	
@@ -39,7 +40,7 @@ extern "C"{
 	Heltal* Heltal_new(int n) {return new Heltal(n);}
 	int Heltal_get(Heltal* heltal) {return heltal->get();}
 	void Heltal_set(Heltal* heltal, int n) {heltal->set(n);}
-	int Heltal_fib(Heltal* heltal) {return heltal->fib();}
+	uint64_t Heltal_fib(Heltal* heltal) {return heltal->fib();}
 	void Heltal_delete(Heltal* heltal){
 		if (heltal){
 			delete heltal;
